+++ Tests/TuistSupportIntegrationTests/System/SystemIntegrationTests.swift
@@ -48,6 +48,23 @@ final class SystemIntegrationTests: TuistTestCase {
         }
     }
 
+    func test_observable_when_it_errors() throws {
+        // Given
+        let observable = subject.observable(["/usr/bin/xcrun", "invalid"]).mapToString()
+
+        // When
+        let result = observable.toBlocking().materialize()
+
+        // Then
+        switch result {
+        case .completed:
+            XCTFail("expected command to fail but it did not")
+        case let .failed(elements, error):
+            XCTAssertTrue(elements.first(where: { $0.value.contains("errno=No such file or directory") }) != nil)
+            XCTAssertEqual(error as? TuistSupport.SystemError, TuistSupport.SystemError.terminated(command: "/usr/bin/xcrun", code: 72))
+        }
+    }
+
     func test_pass_DEVELOPER_DIR() throws {
         try sandbox("DEVELOPER_DIR", value: "/Applications/Xcode/Xcode-10.2.1.app/Contents/Developer/") {
             let result = try subject.capture("env")