+++ Tests/TuistSupportTests/Utils/URLSessionSchedulerTests.swift
@@ -5,10 +5,13 @@ import XCTest
 final class URLSessionSchedulerErrorTests: TuistUnitTestCase {
     func test_type_when_noData() {
         // Given
-        let request = URLRequest(url: URL.test())
+        let url = URL.test()
+        let request = URLRequest(url: url)
+        let status = HTTPStatusCode.notFound
+        let response = HTTPURLResponse(url: url, statusCode: status, httpVersion: nil, headerFields: [:])!
 
         // When
-        let got = URLSessionSchedulerError.noData(request).type
+        let got = URLSessionSchedulerError.httpError(status: status, response: response, request: request).type
 
         // Then
         XCTAssertEqual(got, .abort)
@@ -16,12 +19,15 @@ final class URLSessionSchedulerErrorTests: TuistUnitTestCase {
 
     func test_description_when_noData() {
         // Given
-        let request = URLRequest(url: URL.test())
+        let url = URL.test()
+        let request = URLRequest(url: url)
+        let status = HTTPStatusCode.notFound
+        let response = HTTPURLResponse(url: url, statusCode: status, httpVersion: nil, headerFields: [:])!
 
         // When
-        let got = URLSessionSchedulerError.noData(request).description
+        let got = URLSessionSchedulerError.httpError(status: status, response: response, request: request).description
 
         // Then
-        XCTAssertEqual(got, "An HTTP request to \(request.url!.absoluteString) returned no data")
+        XCTAssertEqual(got, "We got an error \(status) from the request \(response.url!) \(request.httpMethod!)")
     }
 }